ptrace04: under construction

+++ 40__c__linux/ptrace04__revertstring-TODO/ptracer.c
@@ -27,7 +27,7 @@
 #include <string.h>
 
 
-const int long_size = sizeof(long long);
+const int long_size = sizeof(long);
 
 
 void get_data(pid_t child, long addr, char *str, int len)
